pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "fastapi[standard]>=0.129.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-slugify>=8.0.4",
    "sqlalchemy>=2.0.46",
    "email-validator>=2.0.0",
    "bcrypt>=5.0.0",
    "pwdlib[argon2,bcrypt]>=0.3.0",
]
